# src/main/resources/swagger.yaml
swagger: '2.0'

schemes:
  - http
  - https

host: localhost:8085
basePath: /api/bytebank/v1

info:
  version: 1.0.0
  title: ByteBank Api
  description: Administra a movimentação bancária do correntista e controla o acesso através de CPF e senha.
  termsOfService: http://localhost:8085/api/bytebank/v1/termo-uso
  contact:
    name: Suporte
    url: http://localhost:8085/api/bytebank/v1/suporte
    email: boaventura19@yahoo.com.br

securityDefinitions:
  BasicAuth:
    type: basic

security:
  - BasicAuth: []

consumes:
  - application/json
produces:
  - application/json

tags:
  - name: Cliente
  - name: Conta
  - name: Extrato
  - name: Transacao

paths:
  '/clientes':
    post:
      tags:
        - Cliente
      summary: Cadastra Cliente
      description: Cadastra um novo cliente
      operationId: cadastraCliente
      parameters:
        - in: body
          name: cliente_
          schema:
            $ref: '#/definitions/Cliente'
      responses:
        201:
          description: Cadastro efetuado com sucesso
          schema:
            $ref: '#/definitions/Cliente'
        400:
          description: Requisição inválida
        500:
          description: Erro interno no servidor
  
    get:
      tags:
        - Cliente
      summary: listar todos cliente
      description: uma lista com todos cliente
      operationId: listaCliente
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Cliente'
        404:
          description: not found
    
  '/clientes/cpf/{cpf}':
    get:
      tags:
        - Cliente
      summary: buscar por cpf
      description: consultar pelo cpf
      operationId: consultaClientePorCpf
      parameters:
        - in: path
          name: cpf
          type: string
          required: true
      responses:
        201:
          description: Cadastro efetuado com sucesso
          schema:
            $ref: '#/definitions/Cliente'
        404:
          description: not found
        500:
          description: Erro interno no servidor    

  '/conta/{agencia}/{numero}/{digito}/saldo':
    get:
      tags:
        - Conta
      summary: Consulta Saldo
      description: Consulta o saldo da conta do cliente.
      operationId: consultaSaldo
      security:
        - BasicAuth: []
      parameters:
        - in: path
          name: agencia
          type: integer
          format: int32
          required: true
        - in: path
          name: numero
          type: integer
          format: int64
          required: true
        - in: path
          name: digito
          type: integer
          format: int32
          required: true
        - in: header
          name: Authorization
          type: string
          required: true
      responses:
        200:
          description: Consulta a saldo efetuada com sucesso
          schema:
            type: number
            format: double
            example: 134.68
        400:
          description: Requisição inválida
        401:
          description: Requisição não autorizada
        500:
          description: Erro no servidor

  '/extrato/{agencia}/{numero}/{digito}':
    get:
      tags:
        - Extrato
      summary: Consulta Extrato
      description: Consulta o extrato da conta do cliente.
      operationId: consultaExtrato
      security:
        - BasicAuth: []
      parameters:
        - in: path
          name: agencia
          type: integer
          format: int32
          required: true
        - in: path
          name: numero
          type: integer
          format: int64
          required: true
        - in: path
          name: digito
          type: integer
          format: int32
          required: true
        - in: header
          name: Authorization
          type: string
          required: true
      responses:
        200:
          description: Consulta de extrato efetuada com sucesso
          schema:
            $ref: '#/definitions/Extrato'
        400:
          description: Requisição inválida
        401:
          description: Requisição não autorizada
        500:
          description: Erro no servidor

  '/transacao':
    post:
      tags:
        - Transacao
      summary: Cadastra Transacao na Conta
      description: Cadastra uma transacao de debito ou deposito na conta.
      operationId: cadastraTransacao
      security:
        - BasicAuth: []
      parameters:
        - in: header
          name: Authorization
          type: string
          required: true
        - in: body
          name: transacao
          schema:
            $ref: '#/definitions/Transacao'
          required: true
        - in: query
          name: tipo
          type: string
          enum:
          - debito
          - deposito
          required: true
      responses:
        '201':
          description: Transação cadastrada com sucesso!
          schema:
            $ref: '#/definitions/Transacao'
        '400':
          description: Requisição inválida.
        '401':
          description: Requisição não autorizada.
        '500':
          description: Erro no servidor.

  '/transacao/{codigo}':
    put:
      tags:
        - Transacao
      summary: Altera a transação
      description: Altera o valor da transação existente.
      operationId: alteraValorTransacao
      security:
        - BasicAuth: []
      parameters:
        - in: path
          name: codigo
          type: integer
          format: int64
          required: true
        - in: query
          name: valor
          type: number
          format: double
          required: true
        - in: header
          name: Authorization
          type: string
          required: true
      responses:
        '204':
          description: Alteração efeturada com sucesso!
        '400':
          description: Requisição inválida.
        '401':
          description: Requisição não autorizada.
        '404':
          description: Registro não encontrado.
        '500':
          description: Erro no servidor.

    delete:
      tags:
        - Transacao
      summary: Exclui a transação
      description: Exclui a transação existente.
      operationId: excluiTransacao
      security:
        - BasicAuth: []
      parameters:
        - in: path
          name: codigo
          required: true
          type: integer
          format: int64
        - in: header
          name: Authorization
          type: string
          required: true
      responses:
        '204':
          description: Exclusão efetuada com sucesso
        '400':
          description: Requisição inválida.
        '401':
          description: Requisição não autorizada.
        '404':
          description: Registro não encontrado.
        '500':
          description: Erro no servidor

definitions:
  Cliente:
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 234
      titular:
        type: string
        example: Janete Silva
      cpf:
        type: string
        minLength: 11
        maxLength: 11
        example: 01234567890
      senha:
        type: string
        format: password
        example: senha123
      contas:
        type: array
        items:
          $ref: '#/definitions/Conta'

  Conta:
    type: object
    properties:
      agencia:
        type: integer
        format: int32
        example: 975
      numero:
        type: integer
        format: int64
        example: 784563
      digito:
        type: integer
        format: int32
        example: 7
      saldo:
        type: number
        format: double
        example: 101.34
      cliente:
        $ref: '#/definitions/Cliente'
      transacoes:
        type: array
        items:
          $ref: '#/definitions/Transacao'

  Transacao:
    type: object
    properties:
      id:
        type: integer
        format: int64
      tipo:
        type: string
        enum:
          - debito
          - deposito
      valor:
        type: number
        format: double
      data:
        type: string
        format: date-time
      conta:
        $ref: '#/definitions/Conta'

  Extrato:
    type: object
    properties:
      transacoes:
        type: array
        items:
          $ref: '#/definitions/Transacao'      